% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchPred.R
\name{batchPred}
\alias{batchPred}
\title{Predict relevant covariates for batch effect adjustment}
\usage{
batchPred(
  input.edata,
  input.covariates.df,
  input.gold.standard,
  threshold = 1e-04,
  cores = 2
)
}
\arguments{
\item{input.edata}{matrix of numeric expression data, where rows are genes
and columns are samples.}

\item{input.covariates.df}{the covariate dataframe. Each covariate is a 
dataframe column that represents known covariate such as batch number, 
age or hidden covariates such as a principle components.}

\item{input.gold.standard}{reference data table consisting of known gene 
associations. The number of true positives and false positives should be 
approximately equal. gene IDs in \code{input.gold.standard} must match 
gene IDs in \code{input.edata}.}

\item{threshold}{numeric. Minimum AUC score improvement required for 
addition to the linear design.}

\item{cores}{integer. Number of cores / threads. Number of threads is the 
primary bottleneck for computing the initial covariate order.}
}
\value{
\code{batchPred} returns a table consisting of with the following 
columns: 
\itemize{
 \item Covariates: covariate, whose adjustment show greatest improvement in 
 AUC.  
 \item LinearModel: Combination of covariates tested. 
 \item AUC: Area under roc curve (AUC) value after batch adjustment with 
 \code{LinearModel}. 
 \item AUCvRaw: difference in AUC value of batch corrected and raw dataset. 
 \item covEffectOnAUC: improvement in AUC value compared to previously 
 tested \code{LinearModel}.
  }
}
\description{
\code{batchPred} predicts relevant batch variables for improving the 
 normalisation and the interpretation of heterogenous datasets.
}
