% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBceF2.R
\name{plotBceF2}
\alias{plotBceF2}
\title{Plot Receiver-Operator curves for evaluating the performance of batch correction}
\usage{
plotBceF2(
  compute_bcef_object,
  line_color = compute_bcef_object[["line.color"]],
  plot_title = compute_bcef_object[["plot.title"]],
  inner_title = "data_desc",
  showLegendBorder = F,
  title_size = 1.2,
  inner_title_size = 0.8
)
}
\arguments{
\item{compute_bcef_object}{A list of all variables needed to plot ROC curve}

\item{line_color}{Character vector of colors (Optional). Must match number of curves plotted.}

\item{plot_title}{Title of the plot (Optional).}

\item{inner_title}{String. Specify text to go inside the plot area (top left).}

\item{showLegendBorder}{If true, draw border around legend. Default=FALSE}

\item{title_size}{Numeric. Specify size of title}

\item{inner_title_size}{Numeric. Specify size of inner title}
}
\value{
\code{plotBCeF2} Plots ROC curves of raw and batch corrected datasets.
 Allows you to compare the performance of different sets of covariates.
}
\description{
\code{refNet} Plots ROC curve using results from \code{computeBceF2}
}
