% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refNetSubsets.R
\name{refNetSubsets}
\alias{refNetSubsets}
\title{Generates non-overlapping subsets from a reference gene co-expression network}
\usage{
refNetSubsets(
  gold_standard_filt,
  seed = 50,
  number_of_subsets = 3,
  sample_size = 4e+05,
  false_positive = 0.025,
  true_positive = 0.5,
  threshold = NULL
)
}
\arguments{
\item{gold_standard_filt}{Data table of gene-gene network
associations. \code{gold_standard_filt} is basically the output of 
\code{createRefGeneNet}. Column 1 and 2 are entrez gene IDs. Column 3 is 
posterior probability with known edges set to 1. While column 4 defines 
whether a gene relation is a true positive (1) or false positive (0).}

\item{seed}{Integer. Seed allows reproduction of a sequence of random 
numbers. Important for reproducibly subsetting the reference network.}

\item{number_of_subsets}{Integer. Specify the number of reference subsets 
to create. It is recommended to run this function once to generate all 
your subsets at once to avoid gene-gene relation overlap between subsets.}

\item{sample_size}{Integer. Specifies the total number of gene relations 
to extract from \code{gold_standard_filt}. Final number of gene relations
will typically be smaller than this number to ensure a similar
number of true and false positives are generated. Default: 400000}

\item{false_positive}{cutoff value for false positive gene associations.
Should be the same cutoff used in \code{createRefGeneNet} or lower. Used 
as a starting point for adjusting the number of false positives and 
negatives. Default: 0.025}

\item{true_positive}{cutoff value for true positive gene associations.
Should be the same cutoff used in \code{createRefGeneNet} or higher. Used
as a starting point for adjusting the number of false positives and 
negatives. Default: 0.5}
}
\value{
A data table or a list of data tables. Generates non-overlapping 
 subsets from a reference gene association network and calibrates the 
 number of true positives and false positives to be approximately equal.
}
\description{
\code{refNetSubsets} generates non-overlapping subsets from a 
reference gene association network.
}
\examples{
gold_standard_filt <- data.table::data.table(entrezID1 = c(6776, 5359,
    9039, 10651, 6627, 3950, 1998, 222950, 54732, 92745, 
    1974, 30814, 29982, 4967, 10234, 9742, 57484, 5818, 
    5590 , 92211)
 , entrezID2 = c(7318, 3601, 1027, 4190, 6427, 22797, 9215, 
    83642, 10388, 1288, 474, 5973, 1735, 374454, 2886, 
    1933, 79029, 6666, 7307, 51285)
 , postProbability = c(0.62357, 0.784499, 0.535758, 0.824298, 0.773425, 
    0.0131272, 0.00365631, 0.0205155, 0.0201044, 0.0212581, 0.0045336, 
    0.0115648, 0.00168017, 0.0164424, 0.00369498, 0.00608776, 0.0177325, 
    0.0152875, 0.0196584, 0.0222284)
 , Confidence = c(1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0)
 )

subset_list <- refNetSubsets( gold_standard_filt
 , seed = 50
 , number_of_subsets = 3
 , sample_size = 5
 , false_positive = 0.019
 , true_positive = 0.5
 ) 
subset_list[2]
}
