% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBCeF2.R
\name{plotBceF2}
\alias{plotBceF2}
\title{Generates roc curve for evaluating the performance of batch correction}
\usage{
plotBceF2(
  input.edata,
  covariates.df.list,
  input.gold.standard,
  plot.title = "Batch Correction Evaluation",
  roc.curve.legend = NULL,
  line.color = NULL
)
}
\arguments{
\item{input.edata}{Matrix of raw expression dataset to be adjusted. Rows 
represent the genes/probes and columns represent samples.}

\item{covariates.df.list}{List of covariate dataframes. Each dataframe in the 
list consists of columns representing known covariates e.g. Age, Sex, or  
unknown covariatessuch as principle components. Rows represent samples.}

\item{input.gold.standard}{Dataframe. A gold standard that includes a gene 
coexpression confidence. First column represents the first gene, the second 
column the second gene and the third columns should be a binary vector where 
1 indicates true associations and 0 indicates false associations. The gene 
IDs in the first two column of \code{input.gold.standard} and the rownames of
\code{input.edata} must be of the same type.}

\item{plot.title}{Title of the plot.}

\item{roc.curve.legend}{Character vector of descriptions for each set of 
covariates to plot. Must be in the same order as the dataframes in 
\code{covariates.df.list}. This modifies the roc curve legend.}

\item{line.color}{Character vector specifying color of roc curves. By default 
sets color for upto 6 roc curves with raw displayed as black. Manually 
specify line.color if plotting more than 6 different sets of covariates, 
i.e. over 6 dataframes in \code{covariates.df.list}.}
}
\value{
\code{plotBceF2} plots roc curves of raw and batch corrected datasets.
 One batch corrected roc curve is plotted for each dataframe specified in 
 \code{covariates.df.list}, allowing you to compare the performance of 
 different sets of covariates
}
\description{
This package corrects for confounders in gene expression datasets using 
 multiple linear regression model and then evaluates the improvement in gene 
 coexpression of using a gold standard co-expression network.
}
