% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBceF2.R
\name{plotBceF2}
\alias{plotBceF2}
\title{Plot Receiver-Operator curves for evaluating the performance of batch correction}
\usage{
plotBceF2(
  compute_bcef_object,
  line.color = compute_bcef_object[["line.color"]],
  plot.title = compute_bcef_object[["plot.title"]],
  showLegendBorder = F
)
}
\arguments{
\item{compute_bcef_object}{A list of all variables needed to plot ROC curve}

\item{line.color}{Character vector of colors (Optional). Must match number of curves plotted.}

\item{plot.title}{Title of the plot (Optional).}

\item{showLegendBorder}{If true, draw border around legend. Default=FALSE}
}
\value{
\code{plotBCeF2} Plots ROC curves of raw and batch corrected datasets.
 Allows you to compare the performance of different sets of covariates.
}
\description{
\code{refNet} Plots ROC curve using results from \code{computeBceF2}
}
